caribic: move denom registry to benchmark command
